+++ src/ViewUser.js
@@ -1,3 +1,4 @@
+import moment from 'moment';
 import { cloneDeep, get, omit, differenceBy, find } from 'lodash';
 import React from 'react';
 import { compose } from 'redux';
@@ -431,6 +432,9 @@ class ViewUser extends React.Component {
     }
 
     const data = omit(user, ['creds', 'proxies', 'sponsors', 'permissions', 'servicePoints', 'preferredServicePoint']);
+    const today = moment().endOf('day');
+
+    data.active = (moment(user.expirationDate).endOf('day').isSameOrAfter(today));
 
     this.props.mutator.selUser.PUT(data).then(() => {
       this.setState({