+++ src/components/ProxyGroup/ProxyEditList/ProxyEditList.js
@@ -163,7 +163,7 @@ class ProxyEditList extends React.Component {
                 }
                 searchButtonStyle="default"
                 selectUser={user => this.onAdd(user)}
-                visibleColumns={['active', 'name', 'patronGroup', 'username', 'barcode']}
+                visibleColumns={['status', 'name', 'patronGroup', 'username', 'barcode']}
                 columnMapping={columnMapping}
                 disableRecordCreation={disableRecordCreation}
               >