+++ CHANGELOG.md
@@ -21,6 +21,11 @@
 * Implement permission for CRUD Manual Patron Block. Ref UIU-674.
 * Provide `totalCount` to loan MCLs.
 * Support circulation v5.0, requiring service-point information on loans. Refs UIU-717.
+* Format the `active` attribute as `Status` on the find-proxy modal. Fixes UIU-726.
+* Add loan policy name to loan details view. Fixes UIU-715.
+* Refresh action details after renew. Fixes UIU-714.
+* Set expired user status to inactive. Fixes UIU-729.
+* Fix relationship created message. Fixes UIU-730.
 
 ## [2.17.0](https://github.com/folio-org/ui-users/tree/v2.17.0) (2018-10-5)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.16.0...v2.17.0)